+++ persistent/Database/Persist/Class.hs
@@ -1,9 +1,23 @@
 {-# LANGUAGE ConstraintKinds #-}
+
+-- | This module exports all of the type classes in @persistent@ for operating
+-- on the database backends.
+--
+-- @persistent@ offers methods that are abstract in the specific @backend@ type.
+-- For SQL databases, this wil be 'Database.Persist.SqlBackend.SqlBackend'.
+-- Other database backends will define their own types.
+--
+-- Methods and functions in this module have examples documented under an
+-- "Example Usage" thing, that you need to click on to expand.
+--
 module Database.Persist.Class
-    ( ToBackendKey (..)
+    (
 
     -- * PersistStore
-    -- |
+    -- | The 'PersistStore', 'PersistStoreRead', and 'PersistStoreWrite' type
+    -- classes are used to define basic operations on the database. A database
+    -- that implements these classes is capable of being used as a simple
+    -- key-value store.
     --
     -- All the examples present here will be explained based on these schemas, datasets and functions:
     --
@@ -29,9 +43,7 @@ module Database.Persist.Class
     -- > +----+-------+-----+
     -- > |  2 | Simon |  41 |
     -- > +----+-------+-----+
-
-    , PersistCore (..)
-    , PersistStore
+      PersistStore
     , PersistStoreRead (..)
     , PersistStoreWrite (..)
     , PersistRecordBackend
@@ -44,7 +56,9 @@ module Database.Persist.Class
     , insertRecord
 
     -- * PersistUnique
-    -- |
+    -- | The 'PersistUnique' type class is relevant for database backends that
+    -- offer uniqueness keys. Uniquenes keys allow us to perform operations like
+    -- 'getBy', 'deleteBy', as well as 'upsert' and 'putMany'.
     --
     -- All the examples present here will be explained based on these two schemas and the dataset:
     --
@@ -104,12 +118,15 @@ module Database.Persist.Class
     , onlyUnique
 
     -- * PersistQuery
+    -- |  The 'PersistQuery' type class allows us to select lists and filter
+    -- database models.  'selectList' is the canonical read operation, and we
+    -- can write 'updateWhere' and 'deleteWhere' to modify based on filters.
+    , selectList
+    , selectKeys
     , PersistQuery
     , PersistQueryRead (..)
     , PersistQueryWrite (..)
     , selectSource
-    , selectKeys
-    , selectList
     , selectKeysList
 
     -- * DeleteCascade
@@ -133,6 +150,13 @@ module Database.Persist.Class
     , BackendCompatible (..)
     , withCompatibleBackend
 
+    -- * PersistCore
+    -- | 'PersistCore' is a type class that defines a default database
+    -- 'BackendKey' type. For SQL databases, this is currently an
+    -- auto-incrementing inteer primary key. For MongoDB, it is the default
+    -- ObjectID.
+    , PersistCore (..)
+    , ToBackendKey (..)
     -- * JSON utilities
     , keyValueEntityToJSON, keyValueEntityFromJSON
     , entityIdToJSON, entityIdFromJSON
