+++ persistent/Database/Persist/Class/PersistQuery.hs
@@ -1,10 +1,10 @@
 {-# LANGUAGE ExplicitForAll #-}
 module Database.Persist.Class.PersistQuery
-    ( PersistQueryRead (..)
+    ( selectList
+    , PersistQueryRead (..)
     , PersistQueryWrite (..)
     , selectSource
     , selectKeys
-    , selectList
     , selectKeysList
     ) where
 
@@ -22,6 +22,12 @@ import Database.Persist.Class.PersistEntity
 class (PersistCore backend, PersistStoreRead backend) => PersistQueryRead backend where
     -- | Get all records matching the given criterion in the specified order.
     -- Returns also the identifiers.
+    --
+    -- NOTE: This function returns an 'Acquire' and a 'ConduitM', which implies
+    -- that it streams from the database. It does not. Please use 'selectList'
+    -- to simplify the code. If you want streaming behavior, consider
+    -- @persistent-pagination@ which efficiently chunks a query into ranges, or
+    -- investigate a backend-specific streaming solution.
     selectSourceRes
            :: (PersistRecordBackend record backend, MonadIO m1, MonadIO m2)
            => [Filter record]
@@ -66,6 +72,11 @@ class (PersistQueryRead backend, PersistStoreWrite backend) => PersistQueryWrite
 
 -- | Get all records matching the given criterion in the specified order.
 -- Returns also the identifiers.
+--
+-- WARNING: This function returns a 'ConduitM', which implies that it streams
+-- the results. It does not stream results on most backends. If you need
+-- streaming, see @persistent-pagination@ for a means of chunking results based
+-- on indexed ranges.
 selectSource
        :: forall record backend m. (PersistQueryRead backend, MonadResource m, PersistRecordBackend record backend, MonadReader backend m)
        => [Filter record]
@@ -78,6 +89,8 @@ selectSource filts opts = do
     release releaseKey
 
 -- | Get the 'Key's of all records matching the given criterion.
+--
+-- For an example, see 'selectList'.
 selectKeys :: forall record backend m. (PersistQueryRead backend, MonadResource m, PersistRecordBackend record backend, MonadReader backend m)
            => [Filter record]
            -> [SelectOpt record]
@@ -88,11 +101,64 @@ selectKeys filts opts = do
     src
     release releaseKey
 
--- | Call 'selectSource' but return the result as a list.
-selectList :: forall record backend m. (MonadIO m, PersistQueryRead backend, PersistRecordBackend record backend)
-           => [Filter record]
-           -> [SelectOpt record]
-           -> ReaderT backend m [Entity record]
+-- | Returns a @['Entity' record]@ corresponding to the filters and options
+-- provided.
+--
+-- Filters are constructed using the operators defined in "Database.Persist"
+-- (and re-exported from "Database.Persist.Sql"). Let's look at some examples:
+--
+-- @
+-- usersWithAgeOver40 :: 'SqlPersistT' 'IO' ['Entity' User]
+-- usersWithAgeOver40 =
+--     'selectList' [UserAge 'Database.Persist.>=.' 40] []
+-- @
+--
+-- If you provide multiple values in the list, the conditions are @AND@ed
+-- together.
+--
+-- @
+-- usersWithAgeBetween30And50 :: 'SqlPersistT' 'IO' ['Entity' User]
+-- usersWithAgeBetween30And50 =
+--      'selectList'
+--          [ UserAge 'Database.Persist.>=.' 30
+--          , UserAge 'Database.Persist.<=.' 50
+--          ]
+--          []
+-- @
+--
+-- The second list contains the 'SelectOpt' for a record.  We can select the
+-- first ten records with 'LimitTo'
+--
+-- @
+-- firstTenUsers =
+--     'selectList' [] ['LimitTo' 10]
+-- @
+--
+-- And we can select the second ten users with 'OffsetBy'.
+--
+-- @
+-- secondTenUsers =
+--     'selectList' [] ['LimitTo' 10, 'OffsetBy' 10]
+-- @
+--
+-- <https://use-the-index-luke.com/sql/partial-results/fetch-next-page Warning that LIMIT/OFFSET is bad for pagination!>
+--
+-- With 'Asc' and 'Desc', we can provide the field we want to sort on. We can
+-- provide multiple sort orders - later ones are used to sort records that are
+-- equal on the first field.
+--
+-- @
+-- newestUsers =
+--     selectList [] ['Desc' UserCreatedAt, 'LimitTo' 10]
+--
+-- oldestUsers =
+--     selectList [] ['Asc' UserCreatedAt, 'LimitTo' 10]
+-- @
+selectList
+    :: forall record backend m. (MonadIO m, PersistQueryRead backend, PersistRecordBackend record backend)
+    => [Filter record]
+    -> [SelectOpt record]
+    -> ReaderT backend m [Entity record]
 selectList filts opts = do
     srcRes <- selectSourceRes filts opts
     liftIO $ with srcRes (\src -> runConduit $ src .| CL.consume)
